+++ webpack/farm_designer/map/__tests__/garden_map_legend_test.tsx
@@ -28,6 +28,7 @@ describe("<GardenMapLegend />", () => {
       dispatch: jest.fn(),
       tzOffset: 0,
       getConfigValue: jest.fn(),
+      imageAgeInfo: { newestDate: "", toOldest: 1 },
     };
   }
